SPLCIE-1802: make tentative index creation visible for bulk import
